requirements.txt
markdown
pymdown-extensions
mdx-wikilink-plus
pyyaml
